+++ mamikos/database/migrations/2019_11_23_500000_create_orders_table.php
@@ -17,9 +17,9 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('kost_id');
-            $table->unsignedBigInteger('room_count');
-            $table->unsignedBigInteger('total_price');
+            $table->unsignedInteger('kost_id');
+            $table->unsignedInteger('room_count');
+            $table->unsignedInteger('total_price');
             $table->enum('status', Order::ALLOWED_STATUS)->default(Order::READY);
             $table->timestamps();
         });
